% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{logn.seci}
\alias{logn.seci}
\title{95\% Std deviation and confidence interval for lognormal}
\usage{
logn.seci(mu, sd)
}
\arguments{
\item{mu}{Normal mean}

\item{sd}{Normal std dev.}
}
\description{
Calculates the standard deviation and 95% confidence interval for a lognormal
 random variable, given the normal mean and standard deviation.
}
\examples{
sigma.beta=0.24
mu.beta= -1.6
logn.seci(mu.beta,sigma.beta)

}
