% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsppfit.r
\name{Palm2mleData}
\alias{Palm2mleData}
\title{Palm2mleData -- Convert data set from Palm format to the format used by the LCE method.}
\usage{
Palm2mleData(points, cameras, d, l, w, b)
}
\arguments{
\item{points}{Set of observations by both observers, measured in distance units from the start of the transect line.}

\item{cameras}{Vector of integers 1 or 2. For each observation points[i], cameras[i] indicates which camera made the observation.}

\item{d}{Length of the transect line, in the same distance units as points.}

\item{l}{Lag (in seconds) between observer 1 and observer 2.}

\item{w}{Half-width of the detection strip in distance units.}

\item{b}{Buffer half-width (distance from the centre line of the detection strip to the edge of the buffer), beyond which the method assumes that no animal can enter the detection strip between passage of the two observers.}
}
\value{
A list in the format required for the first argument of the segfit function.
}
\description{
Palm2mleData -- Convert data set from Palm format to the format used by the LCE method.
}
